inkless-server/pyproject.toml
[project]
name = "inkless-server"
version = "0.1.0"
description = "AI proxy and automation server for the Inkless receipt printer"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "anthropic>=0.40.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[project.scripts]
inkless-server = "server:main"
